pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Daniel Reichenpfader", email = "daniel.reichenpfader@bfh.ch"},
]
dependencies = [
    "openai>=1.12.0",
    "python-dotenv==1.0.0",
    "streamlit>=1.31.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
run-app = {cmd= "streamlit run ./src/streamlit/Chat.py"}
